requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
tiktoken>=0.7.0
tavily-python>=0.5.0
pypdf>=4.0.0
unstructured>=0.15.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
